Allow HelixClusterManager to adopt different implementation of ClusterChangeHandler
1. Introduce ClusterChangeHandler interface which allows different
implementations
2. Move current ClusterChangeHandler to a separate class
(SimpleClusterChangeHandler)
3. Rewrite some APIs in HelixClusterManager to aggregate cluster-wide
infos from all ClusterChangeHandlers
